requirements.txt
# Production dependencies (pinned for reproducibility)
pynacl==1.6.2
